build(docker): set up multistage build
Separates source aquisition and installation images. This leads to
a smaller final image, and easier customization for those who wish
to use Docker with their own personal forks, or for testing purposes.

For now, defaults to Python 3 environment, but can be easily change to
use Python 2 if a user needs to based on modules.

Also, fixes #2.