feat: add command to change uid/gid for sopel user
Since the container is started as `root`, now we can execute scrips that require privileges. One use of such a script is to change the `uid`/`gid` of the user that runs `sopel`. This is useful to be fix permissions issues caused by mounting configuration files and folders into the container, since they will likely not have the correct permissions initially. `sopel` runs with the user `sopel`, `uid` and `gid` = `100000`. Docker users' `*id`s will not go that high. They will either need to change permissions on the folder, OR, use this script to change the default `*ids` of the `sopel` user.

Fixes 2 of 3 from #9