feat: add ability to change uid/gid for user sopel on the fly
Since the container is started as `root`, now we can execute commands that require privileges. One use of that ability is to change the `uid`/`gid` of the user that runs `sopel` at startup.

The `ENTRYPOINT` now looks for environment variables `PUID` and `PGID` to set the user `sopel` `uid` and `gid`, respectively, before any other commands are executed. This will make it easier to handle permissions for mounted volumes.

Fixes 2 of 3 for #9