fix: install pypi packages one at a time
Install packages sequentially to ensure that dependencies are resolved correctly. Generally, this shouldn't be a problem, but arose due to the issue described in #17.

A side-effect of this fix is that order of package installation can be enforced to be in order of appearance in `/pypi_packages.txt` file.

Fixes #17