build: use travis to autobuild various combinations of Sopel+Python
Added a python script to build and push multiple versions of Sopel for a
given Python-alpine version. travis runs this script for multiple Python
versions. Sopel versions to build are read from `version-map.txt`, a pip
requirements-like text file. Details of the format are found in the
build script, `.travis/build-tag-deploy.py` in the `parse_version_map`
method docstring.