fix: existing .pid file on startup
Prevents the `There's already a Sopel instance running with this config
file` error on startup when a previous (unclean) stop of a docker
container did not remove the old `.pid` file.

Just in case a user has multiple `.pid` files in the same folder (shared
volume mount), the container will still fail if more than one `.pid` is
found.